zooProject Draft
This program utilizes the following Data Structures: 
A Map to keep track of species counts using {Species:Count} Key Value Pairs.
A vector of Animal Objects which store string Age, int Name, and string Species data.
The parsing aspect of the program is currently unfinished, however the file output function should work as soon as the parsing section is completed.
Currently I am very satisfied with how the output function came out. 
The Keys in the Map are iterated through to produce the counts as well as provide the object to create a comparison when looping through the animal object vector.
When on 1 Key the Animal Object Vector is then looped and a getspecies method is called to pull the string to compare to the current key. If a match is found a new line is appended to the output file until all lines have been iterated through. Then the Next Key is check beginning the process once again.
The Parsing section will eventually be executed where each line will be split by spaces, and appended to a vector.
Using a brute force method each data point will be in a consistent spot allowing for the data to be pulled directly like a hashmap. 
Once the parsing data is pulled, the Animal Object creator will be used to create a new object and appended to the Zoo_Animals vector.